+++ Tests/SentryTests/SentryCrash/SentryCrashInstallationReporterTests.swift
@@ -8,7 +8,7 @@ class SentryCrashInstallationReporterTests: XCTestCase {
     private var sut: SentryCrashInstallationReporter!
     
     override func setUp() {
-        sut = SentryCrashInstallationReporter()
+        sut = SentryCrashInstallationReporter(frameInAppLogic: SentryFrameInAppLogic(inAppIncludes: [], inAppExcludes: []))
         sut.install()
         // Works only if SentryCrash is installed
         sentrycrash_deleteAllReports()