+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased
 
+- fix: Mark frames as inApp #956
+
 ## 6.2.0
 
 With this version, Sentry groups errors by domain and code. MyDomain 1 and MyDomain 2